+++ plots/plotsRobert/cutFlowTable.py
@@ -1,6 +1,6 @@
 import ROOT
 
-from StopsDilepton.samples.cmgTuples_Fall15_mAODv2_25ns_1l_postProcessed import *
+from StopsDilepton.samples.cmgTuples_Fall15_mAODv2_25ns_2l_postProcessed import *
 #from StopsDilepton.samples.cmgTuples_FastSimT2tt_mAODv2_25ns_1l_postProcessed import *
 
 from StopsDilepton.tools.helpers import getVarValue, getYieldFromChain
